all-releases/release-list component name and directory changes
all-releases component changed to release-container and moved into release directory

release-list moved into release directory, and into it's own directory called release-list
