--- src/app/music/all-releases.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component ({
-    selector: 'all-releases',
-    templateUrl: 'all-releases.component.html',
-    styleUrls: ['all-releases.component.css']
-})
-
-export class AllReleasesComponent {
-
-}
