+++ src/app/music/releases/release-container.component.html
@@ -1,0 +1,10 @@
+<div class="pageContent">
+    <div class="pageHeader">
+        <h2>Releases</h2>
+    </div>
+    <div class="pageSpan-grayBG">
+        <div class="pageElement">
+            <release-list></release-list>
+        </div>
+    </div>
+</div>
