+++ src/app/music/releases/release-container.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component ({
+    selector: 'release-container',
+    templateUrl: 'release-container.component.html',
+    styleUrls: ['release-container.component.css']
+})
+
+export class ReleaseContainerComponent {
+
+}
