--- src/app/music/releases/release-list.component.css
@@ -1,87 +1,0 @@
-.wrapper {
-    background-color: rgb(240, 240, 240) !important;
-}
-
-.releaseListContainer {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-
-.releaseCardElement {
-    flex-basis: 30%;
-    box-shadow: 0 5px 6px 0 hsla(0, 0%, 0%, 0.3);
-    text-align: center;
-    background-color: white;
-    margin: 0.5em;
-    transition: background-color 0.2s ease;
-}
-
-.releaseCardElement:hover {
-    background-color: #0193b728;
-    cursor: pointer;
-}
-
-.releaseDetails {
-    text-align: left !important;
-    padding: 10px;
-}
-
-.cardHeader {
-    margin-top: 0;
-    margin-bottom: 0;
-    padding-bottom: 10px;
-    padding-top: 10px;
-    color: white;
-    font-weight: 500;
-    background-color: #0193b7;
-}
-
-.title {
-    margin-top: 1px;
-    margin-bottom: 1px;
-    color: rgb(65, 65, 65);
-    font-weight: 800;
-}
-
-.catNo {
-    margin-bottom: 1px;
-    margin-top: 0;
-    color: rgb(65, 65, 65);
-}
-
-.artistName {
-    color: rgb(121, 121, 121);
-    margin-bottom: 0;
-    margin-top: 0;
-}
-
-.artwork {
-    width: 100%;
-    display: block;
-}
-
-.viewAllReleases-button {
-    margin-left: 50px;
-    margin-top: 2px;
-    padding: 5px;
-    background-color: #0193b7;
-    box-shadow: 0 3px 6px 0 hsla(0, 0%, 0%, 0.3);
-    border: none;
-    border-radius: 5px;
-    color: white;
-    text-decoration: none;
-    text-align: center;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-    outline: none;
-    width: 90px;
-    height: 30px;
-  }
-  
-  .viewAllReleases-button:hover {
-    background-color: #00a1c9;   
-  }
