+++ src/app/routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { MainComponent } from './main/main.component';
 import { MusicMainComponent } from './music/music-main.component';
-import { AllReleasesComponent } from './music/all-releases.component';
+import { ReleaseContainerComponent } from './music/releases/release-container.component';
 import { GTA19Component } from './music/releases/release-details/gta019/gta019.component';
 import { GTA18Component } from './music/releases/release-details/gta018/gta018.component';
 import { GTA17Component } from './music/releases/release-details/gta017/gta017.component';
@@ -10,7 +10,7 @@ export const appRoutes:Routes = [
     { path: '', component: MainComponent },
     { path: 'home', component: MainComponent },
     { path: 'music', component: MusicMainComponent},
-    { path: 'releases', component: AllReleasesComponent},
+    { path: 'releases', component: ReleaseContainerComponent},
 
     { path: 'releases/gta19', component: GTA19Component },
     { path: 'releases/gta18', component: GTA18Component },
